add argument init

+++ activetigger/server.py
@@ -43,8 +43,12 @@ class Server():
         with open('config.yaml') as f:
             config = yaml.safe_load(f)
         self.path = Path(config["path"])
-        self.path_fastext = Path(config["path_fasttext"])
         self.SECRET_KEY = config["secret_key"] #generate it automatically ?
+        if config["path_fasttext"] is not None:
+            self.path_fastext = Path(config["path_fasttext"])
+        else:
+            self.path_fastext = None
+            print("Fasttext path model not specified")
 
         # create the database
         self.db = self.path / self.db_name
@@ -59,12 +63,13 @@ class Server():
         # update users from YAML config file
         existing = self.existing_users()
         current = config["users"]
-        for u in current:
-            if not u in existing:
-                self.add_user(u, current[u])
-        for u in existing:
-            if (not u in current) and u != self.default_user:
-                self.delete_user(u)
+        if current is not None:
+            for u in current:
+                if not u in existing:
+                    self.add_user(u, current[u])
+            for u in existing:
+                if (not u in current) and u != self.default_user:
+                    self.delete_user(u)
                 
     def recreate_executor(self):
         """
